+++ NEWS.md
@@ -5,9 +5,7 @@
 * `@code` slot in `qenv` object is now a `character` (previously it was `expression`)
 * `get_code()` is extended by `names` parameter that allows to extract the code just for a
 specific object
-* you can now specify `# @effect object_name` comment tag at the end of the line for the `character` code input in
-`new_qenv()` and `eval_code()` to specify lines having side-effects on objects, so that they are also returned in 
-`get_code()` (when `names` is used)
+* Introduced `# @effect` comment-tag in `code` argument of `eval_code` and `new_qenv` functions to support code relationship detection. See more in functions documentation.
 
 # teal.code 0.4.1